Split up ConfigAlert into multiple files
