+++ src/components/config/alert/ConfigAlertPagerTree.vue
@@ -1,0 +1,89 @@
+<template>
+  <div>
+    <praeco-form-item label="Integration URL" prop="pagertreeIntegrationUrl" required>
+      <el-input id="pagertreeIntegrationUrl" v-model="pagertreeIntegrationUrl" :disabled="viewOnly" />
+      <label>URL generated by PagerTree for the integration.</label>
+    </praeco-form-item>
+
+    <praeco-form-item label="Proxy" prop="pagertreeProxy" required>
+      <el-input id="pagertreeProxy" v-model="pagertreeProxy" :disabled="viewOnly" />
+      <label>
+        By default ElastAlert 2 will not use a network proxy to send notifications to PagerTree.
+        Set this option using hostname:port if you need to use a proxy.
+      </label>
+    </praeco-form-item>
+  </div>
+</template>
+
+<script>
+import validUrl from 'valid-url';
+
+let validateUrl = (rule, value, callback) => {
+  if (validUrl.isUri(value)) {
+    try {
+      let url = new URL(value);
+      if (['http:', 'https:'].includes(url.protocol)) {
+        callback();
+      } else {
+        callback(new Error('Invalid URL'));
+      }
+    } catch (error) {
+      callback(new Error('Invalid URL'));
+    }
+  } else {
+    callback(new Error('Invalid URL'));
+  }
+};
+
+export default {
+  components: {
+  },
+
+  props: ['viewOnly'],
+
+  data() {
+    return {
+      rules: {
+        pagertreeIntegrationUrl: [
+          {
+            validator: validateUrl,
+            trigger: ['change', 'blur']
+          }
+        ]
+      }
+    };
+  },
+
+  computed: {
+    pagertreeIntegrationUrl: {
+      get() {
+        return this.$store.state.config.alert.pagertreeIntegrationUrl;
+      },
+      set(value) {
+        this.$store.commit(
+          'config/alert/UPDATE_PAGERTREE_INTEGRATION_URL',
+          value
+        );
+      }
+    },
+
+    pagertreeProxy: {
+      get() {
+        return this.$store.state.config.alert.pagertreeProxy;
+      },
+      set(value) {
+        this.$store.commit(
+          'config/alert/UPDATE_PAGERTREE_PAGERTREE_PROXY',
+          value
+        );
+      }
+    }
+  },
+
+  methods: {
+  }
+};
+</script>
+
+<style lang="scss">
+</style>
