+++ src/registration.js
@@ -14,6 +14,32 @@ import ElastalertTimeView from '@/components/ElastalertTimeView';
 import ConfigQuery from '@/components/config/ConfigQuery.vue';
 import ConfigAggregation from '@/components/config/ConfigAggregation.vue';
 import ConfigAlert from '@/components/config/alert/ConfigAlert.vue';
+import ConfigAlertAlerta from '@/components/config/alert/ConfigAlertAlerta.vue';
+import ConfigAlertAmazonSes from '@/components/config/alert/ConfigAlertAmazonSes.vue';
+import ConfigAlertAmazonSns from '@/components/config/alert/ConfigAlertAmazonSns.vue';
+import ConfigAlertChatwork from '@/components/config/alert/ConfigAlertChatwork.vue';
+import ConfigAlertCommand from '@/components/config/alert/ConfigAlertCommand.vue';
+import ConfigAlertDatadog from '@/components/config/alert/ConfigAlertDatadog.vue';
+import ConfigAlertDiscord from '@/components/config/alert/ConfigAlertDiscord.vue';
+import ConfigAlertEmail from '@/components/config/alert/ConfigAlertEmail.vue';
+import ConfigAlertExotel from '@/components/config/alert/ConfigAlertExotel.vue';
+import ConfigAlertGitter from '@/components/config/alert/ConfigAlertGitter.vue';
+import ConfigAlertGoogleChat from '@/components/config/alert/ConfigAlertGoogleChat.vue';
+import ConfigAlertHttpPost from '@/components/config/alert/ConfigAlertHttpPost.vue';
+import ConfigAlertJira from '@/components/config/alert/ConfigAlertJira.vue';
+import ConfigAlertLineNotify from '@/components/config/alert/ConfigAlertLineNotify.vue';
+import ConfigAlertMattermost from '@/components/config/alert/ConfigAlertMattermost.vue';
+import ConfigAlertMsTeams from '@/components/config/alert/ConfigAlertMsTeams.vue';
+import ConfigAlertPagerTree from '@/components/config/alert/ConfigAlertPagerTree.vue';
+import ConfigAlertRocketChat from '@/components/config/alert/ConfigAlertRocketChat.vue';
+import ConfigAlertServiceNow from '@/components/config/alert/ConfigAlertServiceNow.vue';
+import ConfigAlertSlack from '@/components/config/alert/ConfigAlertSlack.vue';
+import ConfigAlertStomp from '@/components/config/alert/ConfigAlertStomp.vue';
+import ConfigAlertTelegram from '@/components/config/alert/ConfigAlertTelegram.vue';
+import ConfigAlertTheHive from '@/components/config/alert/ConfigAlertTheHive.vue';
+import ConfigAlertTwilio from '@/components/config/alert/ConfigAlertTwilio.vue';
+import ConfigAlertVictorOps from '@/components/config/alert/ConfigAlertVictorOps.vue';
+import ConfigAlertZabbix from '@/components/config/alert/ConfigAlertZabbix.vue';
 import ConfigKibanaDiscover from '@/components/config/ConfigKibanaDiscover.vue';
 import ConfigTimeWindowFeature from '@/components/config/ConfigTimeWindowFeature.vue';
 import ConfigOwner from '@/components/config/ConfigOwner.vue';
@@ -27,6 +53,32 @@ Vue.component('Bulb', Bulb);
 Vue.component('DateTime', DateTime);
 Vue.component('ConfigQuery', ConfigQuery);
 Vue.component('ConfigAlert', ConfigAlert);
+Vue.component('ConfigAlertAlerta', ConfigAlertAlerta);
+Vue.component('ConfigAlertAmazonSes', ConfigAlertAmazonSes);
+Vue.component('ConfigAlertAmazonSns', ConfigAlertAmazonSns);
+Vue.component('ConfigAlertChatwork', ConfigAlertChatwork);
+Vue.component('ConfigAlertCommand', ConfigAlertCommand);
+Vue.component('ConfigAlertDatadog', ConfigAlertDatadog);
+Vue.component('ConfigAlertDiscord', ConfigAlertDiscord);
+Vue.component('ConfigAlertEmail', ConfigAlertEmail);
+Vue.component('ConfigAlertExotel', ConfigAlertExotel);
+Vue.component('ConfigAlertGitter', ConfigAlertGitter);
+Vue.component('ConfigAlertGoogleChat', ConfigAlertGoogleChat);
+Vue.component('ConfigAlertHttpPost', ConfigAlertHttpPost);
+Vue.component('ConfigAlertJira', ConfigAlertJira);
+Vue.component('ConfigAlertLineNotify', ConfigAlertLineNotify);
+Vue.component('ConfigAlertMattermost', ConfigAlertMattermost);
+Vue.component('ConfigAlertMsTeams', ConfigAlertMsTeams);
+Vue.component('ConfigAlertPagerTree', ConfigAlertPagerTree);
+Vue.component('ConfigAlertRocketChat', ConfigAlertRocketChat);
+Vue.component('ConfigAlertServiceNow', ConfigAlertServiceNow);
+Vue.component('ConfigAlertSlack', ConfigAlertSlack);
+Vue.component('ConfigAlertStomp', ConfigAlertStomp);
+Vue.component('ConfigAlertTelegram', ConfigAlertTelegram);
+Vue.component('ConfigAlertTheHive', ConfigAlertTheHive);
+Vue.component('ConfigAlertTwilio', ConfigAlertTwilio);
+Vue.component('ConfigAlertVictorOps', ConfigAlertVictorOps);
+Vue.component('ConfigAlertZabbix', ConfigAlertZabbix);
 Vue.component('ConfigAggregation', ConfigAggregation);
 Vue.component('ConfigSettings', ConfigSettings);
 Vue.component('ConfigKibanaDiscover', ConfigKibanaDiscover);
